% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/process.KDB.R
\name{process.KDB}
\alias{process.KDB}
\title{Process data from the KDB hierarchy}
\usage{
process.KDB(KDB, remove_initial_spaces = TRUE, extract_tags = TRUE,
  make_lx_ids = TRUE, remove_emptylines = FALSE)
}
\arguments{
\item{KDB}{The KDB data frame returned from read.KDB()}

\item{remove_initial_spaces}{Remove space indents from the content data}

\item{extract_tags}{Extract dictionary codes into a 'tag' column}

\item{make_lx_ids}{Assign a unique sequential integer to each lx item in the dictionary data}

\item{remove_empty_lines}{Remove lines with only spaces. Default is FALSE for retaining human readibility. Suggest TRUE when doing processing to down size the data frame.}
}
\description{
This function performs a number of cleaning and organisation routines to data from KDB.txt. Indiciate whether a routine should or not be performed by adjusting the TRUE/FALSE flags below.
}
\examples{
process.KDB(read.KDB())

KDB <- read.KDB(process_data = FALSE)
KDB <- process.KDB(KDB)

KDB <- process.KDB(KDB, remove_emptylines = TRUE)
}

