% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/read.KDB.R
\name{read.KDB}
\alias{read.KDB}
\title{Read in data from KDB.txt}
\usage{
read.KDB(file_loc = "~/Documents/GitHub/KDB/KDB.txt", process_data = TRUE)
}
\arguments{
\item{file_loc}{Specify where KDB.txt is, this defaults to ~/Documents/GitHub/KDB/KDB.txt}
}
\description{
This function reads in the data from KDB.txt and processes them into a data frame.
}
\examples{
read.KDB()
KDB <- read.KDB()
read.KDB("~/Desktop/KDB_copy.txt")
}
\keyword{cats}

