% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/write.KDB.R
\name{write.KDB}
\alias{write.KDB}
\title{Write KDB data frame to text file}
\usage{
write.KDB(file_loc, df = KDB)
}
\arguments{
\item{file_loc}{Specify the text file to be written. Defaults should be ~/Documents/GitHub/KDB/KDB.txt but will leave this unspecified so the main file does not accidently get overwritten.}

\item{df}{Name of the KDB data frame, assumed to be KDB unless specified.}
}
\description{
This function takes a KDB structured data frame in R and writes the data to a flat file.
}
\examples{
write.KDB(file_loc = "~/Documents/GitHub/KDB/KDB.txt")
}

